% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{as_tibble.dpmaccount_results}
\alias{as_tibble.dpmaccount_results}
\title{Coerce a results object to a tibble}
\usage{
\method{as_tibble}{dpmaccount_results}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"dpmaccount_results"}.}

\item{...}{Additional arguments passed to
\code{\link[tibble:tibble]{tibble::tibble()}}.}
}
\value{
A \link[tibble:tibble]{tibble}.
}
\description{
Convert an object of class \code{"dpmaccount_results"} into
a \link[tibble:tibble]{tibble}.
}
\examples{
df <- as_tibble(results_greenland)
}
\seealso{
An object of class \code{"dpmaccount_results"}
is typically created by a call to \code{\link[=estimate_account]{estimate_account()}}.
}
