% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-helpers.R
\name{augment_events}
\alias{augment_events}
\title{Extract merged data and estimates for events}
\usage{
augment_events(
  object,
  collapse = NULL,
  n_draw = 1000,
  width = 0.95,
  na_rm = FALSE
)
}
\arguments{
\item{object}{An object of class \code{"dpmaccount_results"},
created by \code{\link[=estimate_account]{estimate_account()}}.}

\item{collapse}{Classification variables
to aggregate over: one or more of \code{"age"}, \code{"sex"},
\code{"cohort"}, and \code{"time"}. If collapse is \code{NULL}
(the default), no collapsing is done.}

\item{n_draw}{Number of draws in posterior
samples. Default is 1000.}

\item{width}{A number between 0 and 1 giving
the width of credible intervals. Default is 0.95.}

\item{na_rm}{Whether to replace \code{NA}s in result
with \code{0}s. Default is \code{FALSE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} with:
\itemize{
\item Reported values from input datasets.
\item Lists columns with \code{n_draw} draws from posterior
distributions
\item Point estimates (posterior means), marked \code{.fitted}.
\item Lower and upper limits of credible intervals
specified by \code{width}, marked \code{.lower} and \code{.upper}.
}
}
\description{
Extract merged input data and estimates for
counts of events from a results object
returned by \code{\link[=estimate_account]{estimate_account()}}.
}
\section{Aggregating classification variables}{


\code{\link[=estimate_account]{estimate_account()}} constructs estimates that
are disaggregated by age, sex, cohort, and time.
Aggregated versions of these estimates
be obtained using the \code{collapse} argument. Setting
\code{collapse = "cohort"}, for instance,
yields estimates that are disaggregated only
by age, sex, and time.
}

\examples{
## use results from an existing model
results <- dpmaccount::results_greenland

## extract counts of events using
## complete classsification with
## age, sex, cohort, and time
augment_events(results)

## extract counts of events using
## simpler classification with
## sex and time
augment_events(results, collapse = c("age", "cohort"))
}
\seealso{
\itemize{
\item \code{\link[=augment_population]{augment_population()}} Extract merged data and estimates for population.
\item \code{\link[=augment_rates]{augment_rates()}} Extract merged priors and estimates for rates.
\item \code{\link[=components.dpmaccount_results]{components()}}
Extract estimates, inputs.
\item \code{\link[=diagnostics]{diagnostics()}} Extract measures of model performance.
}
}
