% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-helpers.R
\name{augment_population}
\alias{augment_population}
\title{Extract merged data and estimates for population}
\usage{
augment_population(
  object,
  collapse = NULL,
  n_draw = 1000,
  width = 0.95,
  na_rm = FALSE
)
}
\arguments{
\item{object}{An object of class \code{"dpmaccount_results"},
created by \code{\link[=estimate_account]{estimate_account()}}.}

\item{collapse}{Classification variables
to aggregate over: one or more of \code{"age"}, \code{"sex"},
and \code{"cohort"}. If collapse is \code{NULL}
(the default), no collapsing is done.}

\item{n_draw}{Number of draws in posterior
samples. Default is 1000.}

\item{width}{A number between 0 and 1 giving
the width of credible intervals. Default is 0.95.}

\item{na_rm}{Whether to replace \code{NA}s in result
with \code{0}s. Default is \code{FALSE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} with:
\itemize{
\item Reported values for population from input datasets.
\item A list column with \code{n_draw} draws from the posterior
distribution for population.
\item Point estimates (posterior means), marked \code{.fitted}.
\item Lower and upper limits of credible intervals
specified by \code{width}, marked \code{.lower} and \code{.upper}.
}
}
\description{
Extract data and estimates for population
from a results object
returned by \code{\link[=estimate_account]{estimate_account()}}.
}
\section{Aggregating classification variables}{


\code{\link[=estimate_account]{estimate_account()}} constructs estimates that
are disaggregated by age, sex, cohort, and time.
Aggregated versions of these estimates
be obtained using the \code{collapse} argument. Setting
\code{collapse = "cohort"}, for instance,
yields estimates that are disaggregated only
by age, sex, and time.

Adding up population
counts at different times does not make sense,
so \code{augment_population()} (unlike \code{\link[=augment_events]{augment_events()}})
does not allow \code{collapse} to include \code{"time"}.
}

\examples{
## use results from an existing model
results <- dpmaccount::results_greenland

## extract population counts based on
## complete classification with
## age, sex, cohort, and time
augment_population(results)

## extract population counts based on
## simpler classification with
## sex and time
augment_population(results, collapse = c("age", "cohort"))
}
\seealso{
\itemize{
\item \code{\link[=augment_events]{augment_events()}} Extract merged data and estimates for events.
\item \code{\link[=augment_rates]{augment_rates()}} Extract merged priors and estimates for rates.
\item \code{\link[=components.dpmaccount_results]{components()}}
Extract estimates, inputs.
\item \code{\link[=diagnostics]{diagnostics()}} Extract measures of model performance.
}
}
