% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-helpers.R
\name{augment_rates}
\alias{augment_rates}
\title{Extract merged priors and estimates for demographic rates}
\usage{
augment_rates(
  object,
  collapse = NULL,
  n_draw = 1000,
  width = 0.95,
  na_rm = FALSE
)
}
\arguments{
\item{object}{An object of class \code{"dpmaccount_results"},
created by \code{\link[=estimate_account]{estimate_account()}}.}

\item{collapse}{Classification variables
to aggregate over: one or more of \code{"age"}, \code{"sex"},
and \code{"cohort"}. If collapse is \code{NULL}
(the default), no collapsing is done.}

\item{n_draw}{Number of draws in posterior
samples. Default is 1000.}

\item{width}{A number between 0 and 1 giving
the width of credible intervals. Default is 0.95.}

\item{na_rm}{Whether to replace \code{NA}s in result
with \code{0}s. Default is \code{FALSE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} with four
groups of variables
\itemize{
\item Classification variables. One or more of \code{"age"}, \verb{"sex", }"cohort", and \code{"time"}
\item Draws from the prior distributions for rates,
marked \code{prior}.
\item Draws from the posterior distribution for rates,
marked \code{post}.
\item Direct estimates of rates, obtained from input data
with no modelling.
}

Prior and posterior values include:
\itemize{
\item A list column with \code{n_draw} draws from the
distribution.
\item Point estimates (means), marked \code{.fitted}.
\item Lower and upper limits of credible intervals
specified by \code{width}, marked \code{.lower} and \code{.upper}.
}
}
\description{
Given a results object returned by \code{\link[=estimate_account]{estimate_account()}},
extract three types of demographic rates:
\itemize{
\item Prior values derived from system models.
\item Posterior values derived from the combination of
system models, data models, and data.
\item Direct estimates, derived from data only.
}
}
\details{
The rates are defined as follows:
\itemize{
\item Births: expected births per person-year
\item Deaths: expected deaths per person-year
\item Ins: expected inward movements per year (\emph{not per person-year})
\item Outs: expected outward movements per person-year
}

Estimates for each series include:
\itemize{
\item A list column with \code{n_draw} draws from posterior
distributions.
\item Point estimates (posterior means), marked \code{.fitted}.
\item Lower and upper limits of credible intervals
specified by \code{width}, marked \code{.lower} and \code{.upper}.
}
}
\section{Aggregating classification variables}{


\code{\link[=estimate_account]{estimate_account()}} constructs estimates that
are disaggregated by age, sex, cohort, and time.
Aggregated versions of these estimates
be obtained using the \code{collapse} argument. Setting
\code{collapse = "cohort"}, for instance,
yields estimates that are disaggregated only
by age, sex, and time.

\code{augment_rates()} does not allow \code{collapse} to
include \code{"time"}. (Information on time is
needed for direct estimates.)
}

\examples{
## use results from an existing model
results <- dpmaccount::results_greenland

## extract rates based on
## complete classsification with
## age, sex, cohort, and time
augment_rates(results)

## extract rates based on
## simpler classification with
## sex and time
augment_rates(results, collapse = c("age", "cohort"))
}
\seealso{
\itemize{
\item \code{\link[=augment_population]{augment_population()}} Extract merged data and estimates
for population.
\item \code{\link[=augment_events]{augment_events()}} Extract merged data and estimates for events.
\item \code{\link[=components.dpmaccount_results]{components()}}
Extract estimates, inputs.
\item \code{\link[=diagnostics]{diagnostics()}} Extract measures of model performance.
}
}
