% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{components.dpmaccount_results}
\alias{components.dpmaccount_results}
\title{Extract estimates or input data}
\usage{
\method{components}{dpmaccount_results}(
  object,
  what = "population",
  collapse = NULL,
  n_draw = 1000,
  width = 0.95,
  na_rm = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{"dpmaccount_results"},
created by \code{\link[=estimate_account]{estimate_account()}}.}

\item{what}{Component(s) to extract. Valid
values are \code{"population"}, \code{"events"},
\code{"exposure"}, \code{"rates"}, \code{"sysmods"}', \code{"datamods"},
"\verb{data_population"}, and \code{"data_events"}.}

\item{collapse}{Classification variables
to aggregate over. One or more of \code{"age"}, \code{"sex"},
\code{"cohort"}, and \code{"time"}. If collapse is \code{NULL}
(the default), then no collapsing is done.}

\item{n_draw}{Number of draws in posterior
samples. Default is 1000.}

\item{width}{A number between 0 and 1 giving
the width of credible intervals. Default is 0.95.}

\item{na_rm}{Whether to replace \code{NA}s in result
with \code{0}s. Default is \code{FALSE}.}

\item{...}{Not currently used.}
}
\value{
A \link[tibble:tibble-package]{tibble},
or named list of tibbles.
If ex-post adjustments were needed to generate
non-negative population counts, then a tibble
called 'propn_adjust' is added to the result,
showing the proportion of draws that were adjusted,
by cohort and sex.

Tibbles include:
\itemize{
\item A list column with \code{n_draw} draws from the
distribution.
\item Point estimates (means), marked \code{.fitted}.
\item Lower and upper limits of credible intervals
specified by \code{width}, marked \code{.lower} and \code{.upper}.
}
}
\description{
Extract estimated counts and rates, or input data,
from a results object created by
\code{\link[=estimate_account]{estimate_account()}}.
}
\details{
Argument \code{what} can take one or more of the
following values:
\itemize{
\item \code{"population"} Estimated population counts.
\item \code{"events"} Estimated counts of births, deaths, ins, outs.
\item \code{"exposure"} Estimated person-years of exposure.
\item \code{"rates"} Estimated rates for births, deaths, ins, outs.
\item \code{"sysmods"} Estimated hyper-parameters from
system models for births, deaths, ins, outs.
\item \code{"datamods"} Estimated parameters from data models.
\item \code{"data_population"} Input data on population.
\item \code{"data_events"} Input data on births, deaths, ins, outs.
}
}
\section{Aggregating classification variables}{


\code{\link[=estimate_account]{estimate_account()}} constructs estimates that
are disaggregated by age, sex, cohort, and time.
Aggregated versions of these estimates
be obtained using the \code{collapse} argument. Setting
\code{collapse = "cohort"}, for instance,
yields estimates that are disaggregated only
by age, sex, and time.
}

\examples{
## use results from an existing model
results <- dpmaccount::results_greenland

## extract a single component
components(results, what = "population")

## extract two components
components(results, what = c("events", "rates"))

## population data is returned separately
## from events data, since population data
## refers to time points, and events data
## to periods
components(results, what = c("data_population", "data_events"))

## results from data models
components(results, what = "datamods")
}
\seealso{
\itemize{
\item \code{\link[=augment_population]{augment_population()}} Extract merged data and estimates for population.
\item \code{\link[=augment_events]{augment_events()}} Extract merged data and estimates for events.
\item \code{\link[=augment_rates]{augment_rates()}} Extract merged priors and estimates for rates.
\item \code{\link[=diagnostics]{diagnostics()}} Extract measures of model performance.
}
}
