% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamod-constructors.R
\name{datamod_exact}
\alias{datamod_exact}
\title{Create a data model for a series treated as error-free}
\usage{
datamod_exact(data, nm_series, nm_data = NULL)
}
\arguments{
\item{data}{A data frame}

\item{nm_series}{Name of the demographic series
that the data refers to.}

\item{nm_data}{Name of the dataset.}
}
\value{
An object of class \code{"dpmaccount_datamod_exact"}.
}
\description{
Create a data model in which the observed data are
treated as error-free and are copied straight into
the account and not subsequently updated.
}
\details{
Data for deaths should cover the entire age range,
from 0 up to the maximum age.

Data for births should only cover the reproductive ages,
ie from approximately age 12 to approximately age 49
The exact cutoffs may depend on the data.
It may sometimes be appropriate to ignore a few births at
very low or very high ages, to avoid very low rates,
which can create numerical problems with the estimates.

\code{datamod_exact} is current the only valid data model for births
and deaths, and cannot be used for
any series other than births and deaths.
}
\section{Data}{


\code{data} is a data frame holding reported counts. It must always
have the following variables:
\itemize{
\item \code{age} Non-negative whole numbers. Must start at 0 in the case
of deaths; can start higher in the case of births.
\item \code{sex} Character or factor. Must include the level \code{"Female"}.
Refers to the child in the case of births, and to the person
dying in the case of deaths.
\item \code{time} Whole numbers. Typically years, but can be
quarters or months.
\item \code{cohort} Birth year. Refers to the mother in the case of
births, and to the person dying in the case of deaths.
\item \code{count} Non-negative numbers, not necessarily integer.
}

\code{data} must include every combination of the classification
variables.
}

\section{nm_data}{


The name of the dataset. If no name is supplied,
then the name of the \code{nm_data} argument
is used instead.
}

\section{nm_series}{


The name of the demographic series that the
data refers to. One of
\itemize{
\item \code{births}
\item \code{deaths}
}
}

\examples{
deaths <- dpmaccount::gl_report_deaths
rep_deaths <- datamod_exact(deaths,
  nm_series = "deaths"
)
rep_deaths
}
