% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamod-constructors.R
\name{datamod_nbinom}
\alias{datamod_nbinom}
\title{Specify a data model based on a negative binomial distribution}
\usage{
datamod_nbinom(
  data,
  ratio = 1,
  disp = 1,
  scale_ratio = 0,
  nm_series,
  nm_data = NULL
)
}
\arguments{
\item{data}{A data frame}

\item{ratio}{A single number, or a data frame. (Default is 1.).}

\item{disp}{A single number, or a data frame with a variable called \code{"disp"}. (Default is 1.)}

\item{scale_ratio}{A non-negative number. (Default is 0).}

\item{nm_series}{Name of the demographic series that the data refers to.}

\item{nm_data}{Name of the dataset.}
}
\value{
An object of class \code{"datamod_nbinom"}.
}
\description{
Create a data model where the relationship between the
reported and true population counts is represented by a negative
binomial distribution. In this model, the negative binomial
is given a mean-dispersion parameterisation. (For a discussion
of the mean-disperion parameterisation, see, for instance,
\href{https://en.wikipedia.org/wiki/Negative_binomial_distribution}{wikipedia}
or \href{https://mc-stan.org/docs/2_20/functions-reference/nbalt.html}{stan docs}).
}
\details{
Let \eqn{y} be the reported count and \eqn{x} the true count.
Let \eqn{r} and \eqn{d} be ratio and dispersion parameters.
Let \eqn{v = 1/d}. When \eqn{d > 0}, the probability mass function
is (when \code{scale_ratio} takes the default value of zero)
\deqn{
Pr(Y=y) = \Gamma(v + y) / (y! \Gamma(v))
    (v / (v + r x))^v
    ((r x) / (v + r x))^y
}
When \eqn{d=0}, the distribution reduces to the Poisson distribution,
so that the probability mass function is
\deqn{
Pr(Y=y) = (1/(y!)) e^(-rx) (rx)^y
}

When \code{scale_ratio} is greater than zero the data model becomes
\deqn{
Pr(Y=y) = \Gamma(v + ye^{\alpha_{c_i}}) / ((ye^{\alpha_{c_i}})! \Gamma(v))
    (v / (v + r x))^v
    ((r x) / (v + r x))^(ye^{\alpha_{c_i}})
}
\deqn{\alpha_c \sim \mathcal{N}(0, A_{\alpha}^2)}
\itemize{
\item \eqn{c_i} is the cohort associated with cell \eqn{i},
\item \eqn{\alpha_c} is multiplier applied to the coverage
ratio for cohort \eqn{c}
}

In the extended model, \eqn{\alpha_c}
is treated as unknown, and is estimated
when \code{\link[=estimate_account]{estimate_account()}} is called.
}
\section{data}{


\code{data} is a data frame holding reported counts. It must always
have the following variables:
\itemize{
\item \code{age} Non-negative whole numbers, starting at 0.
\item \code{sex} Character or factor. Must include the level \code{"Female"}.
\item \code{time} Whole numbers. Typically years, but can be
quarters or months.
\item \code{count} Non-negative numbers, not necessarily integer.
}

If the data being modelled describe ins or outs,
then \code{data} must also include a variable
called \code{cohort}, which uses the same units as \code{time}.
If the data being modelled describe population,
then \code{data} must not have a variable called cohort.

If a combination of classifying variables is missing from
\code{data}, then \code{datamod_nbinom} assumes that the corresponding
value for \code{count} is \code{NA} (not \code{0}).
}

\section{disp}{

The \code{disp} argument governs dispersion, ie variability.
When \code{disp} equals 0, variance in the reported
count equals the expected value for the reported count.
When \code{disp} is greater than 0, variance in the reported
count is greater than the expected value. In general,
the less reliable the data source is,
the higher \code{disp} should be.

The \code{disp} argument can be a single number,
or can be a data frame with some or all of
same variables as \code{data},
except for \code{"count"},
which is replaced by a \code{"disp"}.
}

\section{scale_ratio}{


Specifying a non-zero value for \code{scale_ratio} adds
flexibility to the data model.
A non-zero value for \code{scale_ratio} implies that we
may have systematically under-estimated or overestimated
coverage ratios.
Setting \code{scale_ratio} to \code{0.1}, for instance, implies that
the coverage ratios implied by \code{ratio} could consistently
understate or overstate coverage ratios, by 10 percent
or even 20-30 percent.
}

\section{ratio}{


The \code{ratio} argument governs net coverage ratios:
the number of reported events or people expected for each
actual event or person. If a data source is unbiased,
then its coverage ratio is 1.

\code{ratio} can be a single number or a data frame.
If \code{ratio} is a single number, then the
same coverage ratio is used for all values of
the data. If \code{ratio} is a data frame,
then different coverage rates are used for
different parts of the data.
If a classifying variable in \code{data} is
omitted from \code{ratio},  then the coverage ratio
is assumed to be constant
across all levels of the omitted variable.
Data frame \code{ratio} must have a column called \code{ratio}.
}

\section{nm_data}{


The name of the dataset. If no name is supplied,
then the name of the \code{data} argument
is used instead.
}

\section{nm_series}{


The name of the demographic series that the
data refers to. One of
\itemize{
\item \code{"population"}
\item \code{"ins"}
\item \code{"outs"}
}
}

\examples{
## A constant coverage ratio (of 90\%)
## across all categories,
## but higher dispersion for young males,
## than for other groups
reg_popn <- dpmaccount::gl_report_popn
disp <- within(reg_popn, {
  rm(count)
  disp <- ifelse(sex == "Male" & age \%in\% 20:29, 0.01, 0.05)
})
varying_disp_dm <- datamod_nbinom(
  data = reg_popn,
  ratio = 0.9,
  disp = disp,
  nm_series = "population"
)
varying_disp_dm

#' ## A constant but likely biased coverage ratio (of 90\%)
## across all categories,
## but higher dispersion for young males,
## than for other groups
reg_popn <- dpmaccount::gl_report_popn
disp <- within(reg_popn, {
  rm(count)
  disp <- ifelse(sex == "Male" & age \%in\% 20:29, 0.01, 0.05)
})
varying_disp_dm <- datamod_nbinom(
  data = reg_popn,
  ratio = 0.9,
  disp = disp,
  scale_ratio = 0.1,
  nm_series = "population"
)
varying_disp_dm

}
