% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamod-constructors.R
\name{datamod_norm}
\alias{datamod_norm}
\title{Specify a data model based on a normal distribution}
\usage{
datamod_norm(
  data,
  ratio = 1,
  sd,
  scale_ratio = 0,
  scale_sd = 0,
  nm_series,
  nm_data = NULL
)
}
\arguments{
\item{data}{A data frame}

\item{ratio}{A single number, or a data frame.
Defaults to 1.}

\item{sd}{A data frame.}

\item{scale_ratio}{A non-negative number.
Default is 0.}

\item{scale_sd}{A non-negative number.
Default is 0.}

\item{nm_series}{Name of the demographic series
that the data refers to.}

\item{nm_data}{Name of the dataset.}
}
\value{
An object of class \code{"dpmaccount_datamod_norm"}.
}
\description{
Create a data model where the relationship between the
reported and true counts is represented by a
normal distribution.
}
\details{
When \code{scale_ratio} and \code{scale_sd} are both \code{0}
(the default), the data model is

\deqn{y_i \sim \mathcal{N}(\rho_i x_i, \sigma_i^2)}

where
\itemize{
\item \eqn{y_i} is the observed count for cell \eqn{i}
\item \eqn{\rho_i} is the net coverage ratio for cell \eqn{i}
\item \eqn{x_i} is the true count for cell \eqn{i}, and
\item \eqn{\sigma_i} is the standard deviation for cell \eqn{i}.
}

When \code{scale_ratio} and \code{scale_sd} are
greater than \code{0},
the data model becomes

\deqn{y_i \sim \mathcal{N}(e^{\alpha_{c_i}} \rho_i x_i, (e^{\beta_{c_i}} \sigma_i)^2)}
\deqn{\alpha_c \sim \mathcal{N}(0, A_{\alpha}^2)}
\deqn{\beta_c \sim \mathcal{N}(0, A_{\beta}^2)}

where
\itemize{
\item \eqn{c_i} is the cohort associated with cell \eqn{i},
\item \eqn{\alpha_c} is multiplier applied to the coverage
ratio for cohort \eqn{c},
\item \eqn{\beta_c} is multiplier applied to the standard deviation
for cohort \eqn{c},
\item \eqn{A_{\alpha}} is \code{scale_ratio}, and
\item \eqn{A_{\beta}} is \code{scale_sd}.
}

In the extended model, \eqn{\alpha_c}
and \eqn{\beta_c} are
treated as unknown, and are estimated
when \code{\link[=estimate_account]{estimate_account()}}
is called.
}
\section{data}{


\code{data} is a data frame holding reported counts. It must always
have the following variables:
\itemize{
\item \code{age} Non-negative whole numbers, starting at 0.
\item \code{sex} Character or factor. Must include the level \code{"Female"}.
\item \code{time} Whole numbers. Typically years, but can be
quarters or months.
\item \code{count} Non-negative numbers, not necessarily integer.
}

If the data being modelled describe ins or outs,
then \code{data} must also include a variable
called \code{cohort}, which uses the same units as \code{time}.
If the data being modelled describe population,
then \code{data} must not have a variable called cohort.

If a combination of classifying variables is missing from
\code{data}, then \code{datamod_norm} assumes that the corresponding
value for \code{count} is \code{NA} (not \code{0}).
}

\section{ratio}{


The \code{ratio} argument governs net coverage ratios:
the number of reported events or people expected for each
actual event or person. If a data source is unbiased,
then its coverage ratio is 1.

\code{ratio} can be a single number or a data frame.
If \code{ratio} is a single number, then the
same coverage ratio is used for all values of
the data. If \code{ratio} is a data frame,
then different coverage rates are used for
different parts of the data.
If a classifying variable in \code{data} is
omitted from \code{ratio},  then the coverage ratio
is assumed to be constant
across all levels of the omitted variable.
Data frame \code{ratio} must have a column called \code{ratio}.
}

\section{sd}{


The \code{sd} argument governs the standard deviation of
variability around the expected value. In general,
the less reliable the data source is,
the higher \code{sd} should be.

Unlike \code{ratio}, \code{sd} can only be a data frame.
It rarely makes sense to use the same value for
\code{sd} across multiple values for the data.

Data frame \code{sd} must have a variable called \code{sd}.
All values for the \code{sd} variable must be positive
(ie greater than zero).
}

\section{scale_ratio, scale_sd}{


Specifying non-zero values for \code{scale_ratio} or
\code{scale_sd} adds flexibility to the data model.
A non-zero value for \code{scale_ratio} implies that we
may have systematically under-estimated or overestimated
coverage ratios.
A non-zero value for \code{scale_sd} implies that
we may have systematically under-estimated or
overestimated uncertainty. Setting
\code{ratio_sd} to \code{0.1}, for instance, implies that
the coverage ratios implied by \code{ratio} could consistently
understate or overstate coverage ratios, by 10 percent
or even 20-30 percent.
}

\section{nm_data}{


The name of the dataset. If no name is supplied,
then the name of the \code{data} argument
is used instead.
}

\section{nm_series}{


The name of the demographic series that the
data refers to. One of
\itemize{
\item \code{"population"}
\item \code{"ins"}
\item \code{"outs"}
}
}

\examples{
reg_popn <- dpmaccount::gl_report_popn
cover_ratio <- dpmaccount::gl_cover_ratio_popn
cover_sd <- dpmaccount::gl_cover_sd_popn

## specify ratio and sd
datamod_norm(
  data = reg_popn,
  ratio = cover_ratio,
  sd = cover_sd,
  nm_series = "population"
)

## assume data source unbiased - use
## default, ratio = 1
datamod_norm(
  data = reg_popn,
  sd = cover_sd,
  nm_series = "population"
)

## non-zero scale_ratio
datamod_norm(
  data = reg_popn,
  sd = cover_sd,
  scale_ratio = 0.1,
  nm_series = "population"
)
}
