% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamod-constructors.R
\name{datamod_t}
\alias{datamod_t}
\title{Specify a data model based on a t distribution}
\usage{
datamod_t(data, df = 4, ratio = 1, scale, nm_series, nm_data = NULL)
}
\arguments{
\item{data}{A data frame}

\item{df}{Degrees of freedom. A positive
real number. Defaults to 4.}

\item{ratio}{A single number, or a data frame.
Defaults to 1.}

\item{scale}{Values governing dispersion
(equivalent to \code{sd} in normal models.)
A data frame with a column called \code{scale}.}

\item{nm_series}{Name of the demographic series
that the data refers to.}

\item{nm_data}{Name of the dataset.}
}
\value{
An object of class \code{"dpmaccount_datamod_t"}.
}
\description{
Create a data model where the relationship between the
reported and true counts is represented by a
\href{https://en.wikipedia.org/wiki/Student\%27s_t-distribution#Location-scale_t-distribution}{t distribution}
with location and scale parameters
}
\details{
\deqn{y_i \sim t(\nu, \rho_i x_i, s_i^2)}

where
\itemize{
\item \eqn{y_i} is the observed count for cell \eqn{i},
\item \eqn{t} is a location-scale t distribution,
\item \eqn{\nu} is the degrees of freedom,
\item \eqn{\rho_i} is the coverage ratio for cell \eqn{i},
\item \eqn{x_i} is the true count for cell \eqn{i}, and
\item \eqn{s_i} is the scale for cell \eqn{i}.
}
}
\section{df}{


Degrees of freedom. A positive number. Non-integer
values are allowed, but \code{Inf} is not. (Setting
\code{df} to \code{Inf} is equivalent to using
\code{\link[=datamod_norm]{datamod_norm()}}.)

A \eqn{t} distribution look like a normal distribution,
except that it has thicker tails. This means that
extremely low values, or extremely high values,
are more common with a \eqn{t} distribution. The
smaller the value of \code{df}, the thicker the tails,
and the more common extreme values become.
}

\section{scale}{


The \code{scale} argument governs dispersion
around the expected value. In general,
the less reliable the data source is,
the higher \code{scale} should be.

Unlike \code{ratio}, \code{scale} must be a data frame.
It rarely makes sense to use the same value for
\code{scale} across multiple values for the data.

Data frame \code{scale} must have a variable called \code{scale}.
All values for the \code{scale} variable must be positive
(ie greater than zero).
}

\section{data}{


\code{data} is a data frame holding reported counts. It must always
have the following variables:
\itemize{
\item \code{age} Non-negative whole numbers, starting at 0.
\item \code{sex} Character or factor. Must include the level \code{"Female"}.
\item \code{time} Whole numbers. Typically years, but can be
quarters or months.
\item \code{count} Non-negative numbers, not necessarily integer.
}

If the data being modelled describe ins or outs,
then \code{data} must also include a variable
called \code{cohort}, which uses the same units as \code{time}.
If the data being modelled describe population,
then \code{data} must not have a variable called cohort.

If a combination of classifying variables is missing from
\code{data}, then \code{datamod_norm} assumes that the corresponding
value for \code{count} is \code{NA} (not \code{0}).
}

\section{ratio}{


The \code{ratio} argument governs net coverage ratios:
the number of reported events or people expected for each
actual event or person. If a data source is unbiased,
then its coverage ratio is 1.

\code{ratio} can be a single number or a data frame.
If \code{ratio} is a single number, then the
same coverage ratio is used for all values of
the data. If \code{ratio} is a data frame,
then different coverage rates are used for
different parts of the data.
If a classifying variable in \code{data} is
omitted from \code{ratio},  then the coverage ratio
is assumed to be constant
across all levels of the omitted variable.
Data frame \code{ratio} must have a column called \code{ratio}.
}

\section{nm_data}{


The name of the dataset. If no name is supplied,
then the name of the \code{data} argument
is used instead.
}

\section{nm_series}{


The name of the demographic series that the
data refers to. One of
\itemize{
\item \code{"population"}
\item \code{"ins"}
\item \code{"outs"}
}
}

\examples{
reg_popn <- dpmaccount::gl_report_popn
scale <- dpmaccount::gl_cover_sd_popn
names(scale)[match("sd", names(scale))] <- "scale"
datamod_t(
  data = reg_popn,
  scale = scale,
  nm_series = "population"
)

}
