% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{diagnostics}
\alias{diagnostics}
\title{Diagnostics on estimation process}
\usage{
diagnostics(object)
}
\arguments{
\item{object}{An object of class \code{"dpmaccount_results"}}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{cohort} Birth cohort.
\item \code{sex} Sex or gender.
\item \code{nlminb_res} Whether the optimiser \code{\link[=nlminb]{nlminb()}}
returned results.
\item \code{nlminb_conv} Whether \code{nlminb()} reached
convergence (given that \code{nlminb_res} is \code{TRUE}).
\item \code{nlminb_msg} Message returned by \code{nlminb()}.
\item \code{logpost} The value for the unnormalised log-posterior
at the maximum.
\item \code{var_nona} Whether the variance matrix
originally returned by \code{nlminb()} had no NAs or Infs
(given that \code{nlminb_cov} is \code{TRUE}).
\item \code{var_pos} Whether the variance matrix
originally returned by \code{nlminb()} is positive
definite (given that \code{var_nonna} is \code{TRUE}).
\item \code{var_nearpos} Whether a non-positive definite
matrix could be corrected via
\code{\link[Matrix:nearPD]{Matrix::nearPD()}} (given that \code{val_pos} is \code{FALSE}).
\item \code{success} Whether the estimation process succeeded.
\code{TRUE} if and only if \code{var_pos} or \code{var_nearpos} is \code{TRUE}.
}
}
\description{
Get diagnostics on the process of
estimating an account. Separate summaries
are presented for each combination of
cohort and sex, since estimation is carried
out separately for each of these combinations.
}
\section{Warning}{

The value \code{logpost} gives the value of the
\emph{unnormalised} log posterior at the maximum.
The unnormalised log posterior is only defined
up to a constant, and cannot be compared
across cohorts. (Comparing the same cohort across
models may be meaningful, depending on the models.)
}

\examples{
## use results from an existing model
results <- dpmaccount::results_greenland

## get diagnostics
diagnostics(results)
}
\seealso{
\itemize{
\item \code{\link[=components.dpmaccount_results]{components()}}
Extract estimates, inputs.
}
}
