% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpmaccount-package.R
\docType{package}
\name{dpmaccount-package}
\alias{dpmaccount}
\alias{dpmaccount-package}
\title{Estimate a demographic account, using TMB}
\description{
Estimate, extend, or forecast demographic accounts. Inference is done using Template Model Builder ('TMB'), which implements Laplace's method.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ONSdigital/dpmaccount}
  \item Report bugs at \url{https://github.com/ONSdigital/dpmaccount/issues/new?assignees=&labels=bug&projects=&template=bug_report_template.md}
}

}
\author{
\strong{Maintainer}: Office for National Statistics \email{demographic.methods@ons.gov.uk}

Authors:
\itemize{
  \item John Bryant
}

Other contributors:
\itemize{
  \item Aidan Metcalfe [contributor]
  \item Andrew Taylor [contributor]
  \item Daniel Ward [contributor]
}

}
\keyword{internal}
