% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_account.R
\name{estimate_account}
\alias{estimate_account}
\title{Estimate a demographic account}
\usage{
estimate_account(
  sysmods,
  datamods = list(),
  keep_adfun = FALSE,
  include_sysmods = c(1, 1, 1, 1),
  seed_in = NULL
)
}
\arguments{
\item{sysmods}{List of system models
for births, deaths, inward movements,
and outward movements. Specified using
function \code{\link[=sysmod]{sysmod()}}.}

\item{datamods}{List of data models for
population, births, deaths, inward movements,
and outward movements.}

\item{keep_adfun}{Whether to keep the functions
created internally by \code{TMB::MakeADFun()}.
For expert use only. Defaults to \code{FALSE}.}

\item{include_sysmods}{Vector of flags specifying which system models
to include in the estimation, in the format c(births, deaths, ins, outs),
e.g. include_sysmods = c(1, 1, 0, 0) to include births and deaths, and
to exclude ins and outs. Defaults to c(1, 1, 1, 1).}

\item{seed_in}{Integer seed used to generate account seeding.
Defaults to NULL to generate a random seeding.}
}
\value{
An object of class \code{"dpmaccount_results"}.
}
\description{
Estimate consistent counts of population,
births, deaths, and migration, and associated
quantities such as demographic rates.
The main inputs are
\itemize{
\item a set of system models describing regularities
in underlying birth, death, and migration rates
\item a set of data models, describing the relationship
between true and reported counts of population and
events.
}
}
\details{
At present the choice of data models is:
\itemize{
\item \code{\link[=datamod_exact]{datamod_exact()}}, which is used only
with births and deaths
\item \code{\link[=datamod_norm]{datamod_norm()}}, which assumes that errors
are normally-distributed, with known variances.
}

More data models will be added in future.
}
\examples{
## create system models
sysmod_births <- sysmod(
  mean = gl_sysmod_mean_births,
  disp = 0.2,
  nm_series = "births"
)
sysmod_deaths <- sysmod(
  mean = gl_sysmod_mean_deaths,
  disp = 0.2,
  nm_series = "deaths"
)
sysmod_ins <- sysmod(
  mean = gl_sysmod_mean_immig,
  disp = 0.2,
  nm_series = "ins"
)
sysmod_outs <- sysmod(
  mean = gl_sysmod_mean_emig,
  disp = 0.2,
  nm_series = "outs"
)
sysmods <- list(
  sysmod_births,
  sysmod_deaths,
  sysmod_ins,
  sysmod_outs
)

## create data models
datamod_popn <- datamod_norm(
  data = gl_report_popn,
  sd = gl_cover_sd_popn,
  nm_series = "population"
)
datamod_births <- datamod_exact(
  data = gl_report_births,
  nm_series = "births"
)
datamod_deaths <- datamod_exact(
  data = gl_report_deaths,
  nm_series = "deaths"
)
datamod_ins <- datamod_norm(
  data = gl_report_immig,
  sd = gl_cover_sd_immig,
  nm_series = "ins"
)
datamod_outs <- datamod_norm(
  data = gl_report_emig,
  sd = gl_cover_sd_emig,
  nm_series = "outs"
)
datamods <- list(
  datamod_popn = datamod_popn,
  datamod_births = datamod_births,
  datamod_deaths = datamod_deaths,
  datamod_ins = datamod_ins,
  datamod_outs = datamod_outs
)

## do estimation
results <- estimate_account(
  sysmods = sysmods,
  datamods = datamods
)

## object holding estimation results
results

## high-level summary
summary(results)
}
\seealso{
Setting up model:
\itemize{
\item \code{\link[=sysmod]{sysmod()}} System models for rates.
\item \code{\link[=datamod_exact]{datamod_exact()}} Data model for error-free data.
\item \code{\link[=datamod_norm]{datamod_norm()}} Data model assuming normally-distributed
measurement and/or sampling errors.
}

Extracting results:
\itemize{
\item \code{\link[=augment_population]{augment_population()}} Data and estimates for population.
\item \code{\link[=augment_events]{augment_events()}} Data and estimates for events.
\item \code{\link[=augment_rates]{augment_rates()}} priors and estimates for rates.
#' - \code{\link[=components.dpmaccount_results]{components()}}
Lower-level access to data and estiamtes.
\item \code{\link[=diagnostics]{diagnostics()}} Measures of model performance
}
}
