% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-datasets.R
\docType{data}
\name{gl_cover_ratio_immig}
\alias{gl_cover_ratio_immig}
\title{Coverage ratios from simulated coverage survey for
Greenland immigration data}
\format{
A data frame with columns
\code{"age"}, \code{"sex"}, and `"ratio"'.
}
\usage{
gl_cover_ratio_immig
}
\description{
Coverage ratios from simulated coverage survey for
Greenland immigration data
}
\seealso{
Calculated from \code{\link{gl_report_immig}}.
}
\keyword{datasets}
