% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-datasets.R
\docType{data}
\name{gl_cover_sd_emig}
\alias{gl_cover_sd_emig}
\title{Standard error for coverage ratios from simulated coverage survey
for Greenland emigration data}
\format{
A data frame with columns
\code{"age"}, \code{"sex"}, and `"sd".
}
\usage{
gl_cover_sd_emig
}
\description{
Standard error for coverage ratios from simulated coverage survey
for Greenland emigration data
}
\seealso{
Calculated from \code{\link{gl_report_emig}}.
}
\keyword{datasets}
