% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-datasets.R
\docType{data}
\name{gl_cover_sd_immig}
\alias{gl_cover_sd_immig}
\title{Standard deviation for coverage ratios from simulated coverage survey
for Greenland immigration data}
\format{
A data frame with columns
\code{"age"}, \code{"sex"}, and `"sd".
}
\usage{
gl_cover_sd_immig
}
\description{
Standard deviation for coverage ratios from simulated coverage survey
for Greenland immigration data
}
\seealso{
Calculated from \code{\link{gl_report_immig}}.
}
\keyword{datasets}
