% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-datasets.R
\docType{data}
\name{gl_cover_sd_popn}
\alias{gl_cover_sd_popn}
\title{Standard deviation for coverage ratios from simulated coverage survey
for Greenland population data}
\format{
A data frame with columns
\code{"age"}, \code{"sex"}, and `"sd".
}
\usage{
gl_cover_sd_popn
}
\description{
Standard deviation for coverage ratios from simulated coverage survey
for Greenland population data
}
\seealso{
Calculated from \code{\link{gl_report_popn}}.
}
\keyword{datasets}
