% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-datasets.R
\docType{data}
\name{gl_report_births}
\alias{gl_report_births}
\title{Registered births in Greenland}
\format{
A data frame with columns \code{"cohort"},
\code{"age"}, \code{"sex"},
\code{"time"}, and \code{"count"}.
}
\source{
The StatsBank database on the Statistics Greenland
website, accessed using package 'pxweb'.
}
\usage{
gl_report_births
}
\description{
Registered births in Greenland by cohort and age of mother,
gender of child, and calendar year.
}
\keyword{datasets}
