% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-datasets.R
\docType{data}
\name{gl_report_emig}
\alias{gl_report_emig}
\title{Registered emigration from Greenland}
\format{
A data frame with columns \code{"cohort"},
\code{"age"}, \code{"sex"},
\code{"time"}, and \code{"count"}.
}
\source{
The StatsBank database on the Statistics Greenland
website, accessed using package 'pxweb'.
}
\usage{
gl_report_emig
}
\description{
Registered emigration from Greenland by cohort, age,
gender, and calendar year.
}
\keyword{datasets}
