% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-datasets.R
\docType{data}
\name{gl_report_popn}
\alias{gl_report_popn}
\title{Registered population in Greenland}
\format{
A data frame with columns \code{"cohort"},
\code{"age"}, \code{"sex"},
\code{"time"}, and \code{"count"}.
}
\source{
The StatsBank database on the Statistics Greenland
website, accessed using package 'pxweb'.
}
\usage{
gl_report_popn
}
\description{
Registered population
in Greenland at the end of the calendar year,
by cohort, age,
sex, and calendar year.
}
\keyword{datasets}
