% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-datasets.R
\docType{data}
\name{gl_sysmod_mean_births}
\alias{gl_sysmod_mean_births}
\title{Estimates of birth rates in Greenland}
\format{
A data frame with columns
\code{"age"}, \code{"time"}, and \code{"mean"}.
}
\source{
Calculated from \code{\link{gl_report_births}} and
\code{\link{gl_report_popn}}.
}
\usage{
gl_sysmod_mean_births
}
\description{
Mean of posterior distribution for
birth rates in Greenland by age of mother
and calendar year.
}
\keyword{datasets}
