% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-datasets.R
\docType{data}
\name{gl_sysmod_mean_deaths}
\alias{gl_sysmod_mean_deaths}
\title{Estimates of death rates in Greenland}
\format{
A data frame with columns
\code{"age"}, \code{"sex"},
\code{"time"}, and \code{"mean"}.
}
\source{
Calculated from \code{\link{gl_report_deaths}}
and
\code{\link{gl_report_popn}}.
}
\usage{
gl_sysmod_mean_deaths
}
\description{
Mean of posterior distribution for
death rates in Greenland by age, sex,
and calendar year.
}
\keyword{datasets}
