% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-datasets.R
\docType{data}
\name{gl_sysmod_mean_immig}
\alias{gl_sysmod_mean_immig}
\title{Estimates of immigration 'rates' in Greenland}
\format{
A data frame with columns
\code{"age"}, \code{"sex"},
\code{"time"}, and \code{"mean"}.
}
\source{
Calculated from \code{\link{gl_report_immig}}.
}
\usage{
gl_sysmod_mean_immig
}
\description{
Mean of posterior distribution for
expected counts for immigration in Greenland by age, sex,
and calendar year.
}
\keyword{datasets}
