% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-helpers.R
\name{increments}
\alias{increments}
\title{Calculate annual increments implied by data}
\usage{
increments(object)
}
\arguments{
\item{object}{An object of class \code{"dpmaccount_results"},
typically created by a call to \code{\link[=estimate_account]{estimate_account()}}.}
}
\value{
A tibble.
}
\description{
Calculate annual increments in population size
implied by input data on stocks and flows.
}
\examples{
results_greenland |>
  increments()
}
