% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codatamod-methods.R
\name{increments_codatamod}
\alias{increments_codatamod}
\title{Calculate increments within Lexis triangles}
\usage{
increments_codatamod(x, cohort, nm_data, is_stock)
}
\arguments{
\item{x}{Object of class "dpmaccount_codatamod"}

\item{cohort}{Numeric - cohort that data refers to}

\item{nm_data}{Name of the dataset}

\item{is_stock}{Whether the object holds stock data}
}
\value{
A tibble
}
\description{
Take time and age information and calculates corresponding cohort and Lexis triangle indexing
}
