% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{post_pred}
\alias{post_pred}
\title{Draw estimates for the posterior predictive distributions}
\usage{
post_pred(
  object,
  what = c("population", "events"),
  width = 0.95,
  n_draw = 1000,
  na_rm = FALSE,
  ...
)
}
\arguments{
\item{object}{the dpmaccount_results object}

\item{what}{Component(s) to extract. Valid
values are \code{"population"}, \code{"events"}}

\item{width}{width of the augment bounds}

\item{n_draw}{numbers of draws to sample from the}

\item{na_rm}{Whether to replace \code{NA}s in result
with \code{0}s. Default is \code{FALSE}.}

\item{...}{Not currently used.}
}
\value{
A \link[tibble:tibble-package]{tibble},
or named list of tibbles.

Tibbles include:
\itemize{
\item A list column with \code{n_draw} draws from the
posterior predictive distribution.
\item Point estimates (means), marked \code{.fitted}.
\item Lower and upper limits of credible intervals
specified by \code{width}, marked \code{.lower} and \code{.upper}.
\item dataset tilde
}
}
\description{
Argument \code{what} can take one or more of the
following values:
\itemize{
\item \code{"population"} Estimated population counts.
\item \code{"events"} Estimated counts of births, deaths, ins, outs.
system models for births, deaths, ins, outs.
}
}
\section{Aggregating classification variables}{


\code{\link[=estimate_account]{estimate_account()}} constructs estimates that
are disaggregated by age, sex, cohort, and time.
}

\examples{
## use results from an existing model
results <- dpmaccount::results_greenland

## posterior predictive
post_pred(results, what = c("events", "population"), n_draw = 2)

}
\seealso{
\itemize{
\item \code{\link[=augment_population]{augment_population()}} Extract merged data and estimates for population.
\item \code{\link[=augment_events]{augment_events()}} Extract merged data and estimates for events.
\item \code{\link[=augment_rates]{augment_rates()}} Extract merged priors and estimates for rates.
\item \code{\link[=diagnostics]{diagnostics()}} Extract measures of model performance.
}
}
