% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comod-methods.R, R/results-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fit}
\alias{as_tibble}
\alias{components}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{components}}, \code{\link[generics]{fit}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

