% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_comod}
\alias{sim_comod}
\title{Simulate a comod object}
\usage{
sim_comod(
  K = 20,
  mean_stk_init = 1000,
  rate_exit = 0.01,
  disp = 0.1,
  sd_relative = 0.1,
  scale_ratio = 0,
  scale_sd = 0,
  is_new_cohort = FALSE
)
}
\arguments{
\item{K}{Number of Lexis triangles}

\item{mean_stk_init}{Prior expected
value for initial stock}

\item{rate_exit}{Death rate and
outs rate (which are assumed to be identical)}

\item{disp}{Overdispersion. Same
value applied to all rates.}

\item{sd_relative}{Size of errors
in (normal) data models}

\item{scale_ratio, scale_sd}{Arguments passed to datamod_norm.}

\item{is_new_cohort}{Whether the cohort
being simulated was born during the
estimation period.}
}
\value{
An object of class "dpmaccount_comod"
}
\description{
Simulate a comod object
}
\section{Warning}{
 The interface for this function
is still evolving: eg we need to include the
ability to specify non-normal data models.
}

\keyword{internal}
