% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comod-methods.R
\name{summary.dpmaccount_comod}
\alias{summary.dpmaccount_comod}
\title{Summarise a cohort model}
\usage{
\method{summary}{dpmaccount_comod}(object, ...)
}
\arguments{
\item{object}{A \code{dpmaccount_comod} object, not necessarily fitted.}

\item{...}{Not currently used.}
}
\value{
An object of class \code{"dpmaccount_comod_summary"}.
}
\description{
Summarise a cohort model
}
\examples{
results <- dpmaccount::results_greenland
fitted <- results$fitted
fitted[[12]]
summary(fitted[[12]])
}
