% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-methods.R
\name{summary.dpmaccount_results}
\alias{summary.dpmaccount_results}
\title{Summarise results from a call to estimate_account}
\usage{
\method{summary}{dpmaccount_results}(object, ...)
}
\arguments{
\item{object}{Object of class \code{"dpmaccount_results"}.}

\item{...}{Currently ignored.}
}
\value{
A named list, invisibly.
}
\description{
Summary method for an object of class
\code{"dpmaccount_results"}, typically created by
a call to \code{\link[=estimate_account]{estimate_account()}}. The summary consists
of diagnostic information, plus estimated counts
and rates for each time period (with no age,
sex, or cohort detail.)
}
