% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysmod-constructors.R
\name{sysmod}
\alias{sysmod}
\title{Specify a system model}
\usage{
sysmod(mean, disp = 0, nm_series)
}
\arguments{
\item{mean}{A data frame.}

\item{disp}{A data frame or a single number.}

\item{nm_series}{One of \code{"births"}, \code{"deaths"},
\code{"ins"}, \code{"outs"}.}
}
\value{
An object of class \code{"dpmaccount_sysmod"}.
}
\description{
Specify a system model for rates for
\itemize{
\item births,
\item deaths,
\item ins (eg inward migration)
\item outs (eg outward migration).
}
}
\details{
Rates for births, deaths, and outs are all expressed
as events per person-year lived, which is
the most common way of defining demographic rates.
However, \strong{rates for ins are expressed as events
per year}, reflecting the fact that there is no
natural population at risk for inward movements.
When specifying rates for \code{ins}, it is important to
check that the rates are in fact defined as events
per year. For instance, if \code{ins} rates are classified by
age, period, and cohort, or by age, period, and Lexis
triangle, then they may have been defined as
events per half-year, rather than events per year.

The model assumes that the rates are drawn from a
gamma distribution with shape parameter equal
to \code{1/disp}, and shape parameter equal to
\code{disp * mean}. When \code{disp} is zero, the rates
are treated as known.

\code{mean} is a data frame with a variable called \code{mean}
containing non-negative numbers, and at least one
of the following variables:
\itemize{
\item \code{age} Non-negative whole numbers, starting at 0
except in the case of births.
\item \code{sex} Character or factor.
\item \code{cohort} Birth cohort. Whole numbers, using the same
units as \code{time}.
\item \code{time} Whole numbers. Normally years, but can be months
or quarters.
}

The classifying variables that do occur in \code{mean} must
contain all categories that appear in the demographic account.

\code{disp} can be a single non-negative number, or a data frame.
If \code{disp} is a data frame, then it is specified in the same
way as \code{mean}, except that the measurement variable is
called \code{disp} rather than \code{mean}.

Rates for births should only cover the reproductive ages,
ie from approximately age 12 to approximately age 49
The exact cutoffs may depend on the births data.
It may sometimes be appropriate to ignore a few births at
very low or very high ages, to avoid very low rates,
which can create numerical problems with the estimates.
}
\section{Multiple draws}{

At present we only allow for single draws
for means and dispersions. We can extend this in
future by allowing for \code{value} columns that
are (for instance) lists or matrices.
}

\examples{
mean <- dpmaccount::gl_sysmod_mean_deaths
sysmod_deaths <- sysmod(
  mean = mean,
  disp = 0.2,
  nm_series = "deaths"
)
sysmod_deaths
}
