% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comod-methods.R
\name{fit.dpmaccount_comod}
\alias{fit.dpmaccount_comod}
\title{Fit a cohort model}
\usage{
\method{fit}{dpmaccount_comod}(object, keep_adfun = FALSE, ...)
}
\arguments{
\item{object}{A \code{dpmaccount_comod} object.}

\item{keep_adfun}{Whether to keep function created
by TMB::MakeADFun.}

\item{...}{Not currently used.}
}
\value{
A fitted \code{dpmaccount_comod} object
}
\description{
For details on the algorithm used by
the optimiser \code{\link[stats:nlminb]{stats::nlminb()}}, see the discussion
\href{here}{https://stackoverflow.com/questions/49375840/algorithm-name-in-nlminbs-port-routines}.
Note that although the online help for \code{nlminb()} suggests
using function \code{\link[stats:optim]{stats::optim()}} instead,
most applications of TMB (eg the glmmTMB package)
seem to use \code{nlminb()}.
}
\section{Warning}{

The optimiser currently fails for some cases.
}

\keyword{internal}
